+++ firmware/lucidgloves-firmware/src/Controller/Gesture.h
@@ -4,9 +4,9 @@
 
 class Gesture {
 public:
-    bool grabGesture(int *flexion);
-    bool pinchGesture(int *flexion);
-    bool triggerGesture(int *flexion);
+    bool grabGesture(float *flexion);
+    bool pinchGesture(float *flexion);
+    bool triggerGesture(float *flexion);
 };
 
 #endif
